release2

+++ SpringBoot/src/main/java/cn/edu/xidian/library/entity/User.java
@@ -18,5 +18,5 @@ public class User {
     @TableField(exist = false)
     private String token;
     private Integer role;
-    private Double debt;
+    private Integer debt;
 }
